lib:sbi: Detect SSCOF extension at run time
SSCOF ISA extension introduces PMU counter overflow and filtering support.
It introduces a read only `scountovf` csr that can be used to detect if
a hart supports this extension at runtime. However, this feature is only
useful if the hart already supports mcounteren and mcountinhibit.

Add a dynamic detection mechanism and boot time print message if sscof
is present.

Signed-off-by: Atish Patra <atish.patra@wdc.com>