pmu: shadow cycle, insret
implement shadow cycle, insret:
- on pmu_ctr_start_fw we store current cycle/insret value
- on pmu_ctr_read_fw we return curr_count + calculated delta

currently reading CSR_MCYCLE, CSR_MINSTRET registers is a problem
in case we also ordered real cycle and insret, because pmu module
will overwrite registers to set it's own provided value.

So if we are using something that relies on reading CSR_MCYCLE and
CSR_MINSTRET and we also need mcycle and minsret - we can use the shadow
version of mentioned hardware counters, as they don't write to registers
on start.

Signed-off-by: Nikita Shubin <n.shubin@yadro.com>