lib: pmu: support the event ID encoded by a bitmap.
RISC-V privilege specification does not specify how to encode the event ID.
Therefore, each platform is allowed to customize its own encoding rule.
The common encoding methods are as follow, directly assigning a number to an
event, or every bit in the mphmevent CSR controls one specified event or
mixes the above two methods.

To enable OpenSBI to support the above three encoding methods simultaneously,
this patch modifies the content of "pmu,raw-event-to-mhpmcounters". The
"pmu,raw-event-to-mhpmcounters" still describes the one or multiple raw
events that could be counted by which mhpmcounters. But, the column number
of "pmu,raw-event-to-mhpmcounters" is extended from 2 to 3. The 1st column
(64bit) is the ID of the raw events. The 2nd column (64bit) is select_mask.
It indicates which bits in the 1st column are encoded by a number. If a
platform directly encodes each PMU event as a different number, in this
case, the value of select_mask will be 0xffffffff_ffffffff. For the other
extreme case, if a platform specified the 64 bits of mhpmevent to control
64 different monitor event, in this case, the value of select_mask will be
zero. The 3rd column(32bit) is the counter map used to describe which
mhpmcounters support the raw events. All the modifications in this patch
are all based on the above change in "pmu,raw-event-to-mhpmcounters".

Signed-off-by: Vincent Chen <vincent.chen@sifive.com>